move "did_propagate_command_buffer" test
