+++ crates/bevy_transform/src/propagate.rs
@@ -28,53 +28,10 @@ impl Propagatable for Transform {
 mod test {
     use bevy_app::prelude::*;
     use bevy_ecs::prelude::*;
-    use bevy_ecs::system::CommandQueue;
     use bevy_math::vec3;
 
     use crate::components::{GlobalTransform, Transform};
-    use crate::TransformBundle;
-    use bevy_hierarchy::{BuildChildren, BuildWorldChildren, Children};
-
-    #[test]
-    fn did_propagate_command_buffer() {
-        let mut world = World::default();
-        let mut update_stage = SystemStage::parallel();
-        update_stage.add_system(bevy_hierarchy::propagate_system::<Transform>);
-
-        let mut schedule = Schedule::default();
-        schedule.add_stage("update", update_stage);
-
-        // Root entity
-        let mut queue = CommandQueue::default();
-        let mut commands = Commands::new(&mut queue, &world);
-        let mut children = Vec::new();
-        commands
-            .spawn_bundle(TransformBundle::from(Transform::from_xyz(1.0, 0.0, 0.0)))
-            .with_children(|parent| {
-                children.push(
-                    parent
-                        .spawn_bundle(TransformBundle::from(Transform::from_xyz(0.0, 2.0, 0.0)))
-                        .id(),
-                );
-                children.push(
-                    parent
-                        .spawn_bundle(TransformBundle::from(Transform::from_xyz(0.0, 0.0, 3.0)))
-                        .id(),
-                );
-            });
-        queue.apply(&mut world);
-        schedule.run(&mut world);
-
-        assert_eq!(
-            *world.get::<GlobalTransform>(children[0]).unwrap(),
-            GlobalTransform::from_xyz(1.0, 0.0, 0.0) * Transform::from_xyz(0.0, 2.0, 0.0)
-        );
-
-        assert_eq!(
-            *world.get::<GlobalTransform>(children[1]).unwrap(),
-            GlobalTransform::from_xyz(1.0, 0.0, 0.0) * Transform::from_xyz(0.0, 0.0, 3.0)
-        );
-    }
+    use bevy_hierarchy::{BuildWorldChildren, Children};
 
     #[test]
     fn correct_transforms_when_no_children() {
